pyproject.toml
[tool.poetry]
name = "arbsense"
version = "0.1.0"
description = ""
authors = ["Ian Draves <iantdraves@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
betiq = "^18"
pandas = "^2.1.0"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
